# server/docs/swagger.yaml
basePath: /api
definitions:
...
  api.changePwdReq:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  api.loginReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  store.Agent:
    properties:
      agent_id:
        type: string
      arch:
        type: string
      created_at:
        type: string
      hostname:
        type: string
      ipv4:
        items:
          $ref: '#/definitions/store.IPAddr'
        type: array
      ipv6:
        items:
          $ref: '#/definitions/store.IPAddr'
        type: array
      kernel:
        type: string
      last_seen:
        type: string
      os:
        type: string
      status:
        description: Status 由 API 层计算（服务端权威）：online | offline，预留 upgrading 等扩展状态。
        type: string
      version:
        description: Agent 构建版本（注册上报）
        type: string
    type: object
...
  store.ContainerSample:
    properties:
      agent_id:
        type: string
      container_id:
        type: string
      cpu_pct:
        type: number
      image:
        type: string
      mem_limit:
        type: integer
      mem_usage:
        type: integer
      name:
        type: string
      net_rx_bps:
        type: integer
      net_tx_bps:
        type: integer
      restarts:
        type: integer
      state:
        type: string
      ts:
        type: string
      uptime_s:
        type: integer
    type: object
  store.DiskSample:
    properties:
      agent_id:
        type: string
      fs:
        type: string
      mount:
        type: string
      total:
        type: integer
      ts:
        type: string
      used:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    name: litesentry
    url: https://github.com/
  description: 轻量主机 + 容器监控 Server REST 接口（前端面板取数）。Agent↔Server 走 gRPC，契约见 docs/proto.html。
  termsOfService: http://localhost:8080/
  title: litesentry API
  version: 0.1.0
paths:
  /agents:
    get:
      description: 全部 Agent 最新状态，含 IPv4/IPv6 地址快照
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Agent'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 节点列表
      tags:
      - 节点
  /agents/{id}/containers:
    get:
      description: 某节点在窗口内各容器的状态与资源样本，按时间升序
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: 起始 unix 秒（默认 1 小时前）
        in: query
        name: from
        type: integer
      - description: 结束 unix 秒（默认当前）
        in: query
        name: to
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.ContainerSample'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 容器指标时序
      tags:
      - 时序
  /agents/{id}/disks:
    get:
      description: 某节点在窗口内各挂载点的磁盘占用样本，按时间升序
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: 起始 unix 秒（默认 1 小时前）
        in: query
        name: from
        type: integer
      - description: 结束 unix 秒（默认当前）
        in: query
        name: to
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.DiskSample'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 磁盘指标时序
      tags:
      - 时序
  /agents/{id}/host:
    get:
      description: 某节点在窗口内的主机样本（CPU/内存/负载/网络/系统信息），按时间升序
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: 起始 unix 秒（默认 1 小时前）
        in: query
        name: from
        type: integer
      - description: 结束 unix 秒（默认当前）
        in: query
        name: to
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.HostSample'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 主机指标时序
      tags:
      - 时序
  /alerts/events:
    get:
      description: 按时间窗口查询，可按节点 / 状态过滤；默认最近 24h
      parameters:
      - description: 起始 unix 秒
        in: query
        name: from
        type: integer
      - description: 结束 unix 秒
        in: query
        name: to
        type: integer
      - description: 节点 ID
        in: query
        name: agent_id
        type: string
      - description: firing | resolved
        in: query
        name: state
        type: string
      - description: 条数（默认 100，最大 500）
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.AlertEvent'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 告警事件列表
      tags:
      - 告警
  /alerts/rules:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.AlertRule'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 告警规则列表
      tags:
      - 告警
    post:
      consumes:
      - application/json
      parameters:
      - description: 规则（id 忽略）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.AlertRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.AlertRule'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 新建告警规则
      tags:
      - 告警
  /alerts/rules/{id}:
    delete:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除告警规则
      tags:
      - 告警
    put:
      consumes:
      - application/json
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: string
      - description: 规则（id 取自路径）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.AlertRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.AlertRule'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新告警规则
      tags:
      - 告警
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: 校验旧密码，写入新密码并清除 must_change_password
      parameters:
      - description: 旧/新密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.changePwdReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 旧密码错误 / 新密码过短
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 修改密码
      tags:
      - 认证
  /auth/login:
    post:
      consumes:
      - application/json
      description: 校验用户名密码，成功签发 JWT 并回写上次登录时间；新用户/被重置密码的用户返回 must_change_password=true
      parameters:
      - description: 用户名密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.loginResp'
        "400":
          description: 参数缺失
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 凭据错误
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: 失败次数过多已锁定
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 登录
      tags:
      - 认证
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
      security:
      - BearerAuth: []
      summary: 当前用户
      tags:
      - 认证
  /frp:
    get:
      description: 全部 frps/frpc 配置，join 最近上报状态与目标节点主机名（token 不回传）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.frpView'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: FRP 配置列表
      tags:
      - FRP
  /frp/{kind}/{agent_id}:
    delete:
      description: 删除 frps/frpc 配置并下线对应 frp 进程（agent 下次心跳生效）
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除 FRP 配置
      tags:
      - FRP
    put:
      consumes:
      - application/json
      description: 新建（agent_id 不存在时）或更新 frps/frpc 配置；token 留空 = 保留原值
      parameters:
      - description: frps | frpc
        in: path
        name: kind
        required: true
        type: string
      - description: 目标 agent（frps 恒为 server）
        in: path
        name: agent_id
        required: true
        type: string
      - description: 配置体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.saveFrpBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.frpView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 保存 FRP 配置
      tags:
      - FRP
...
  /overview:
    get:
      description: 一次返回窗口内每节点最新主机样本、每节点最高磁盘占用、当前 firing 事件数
      parameters:
      - description: 起始 unix 秒（默认 1 小时前）
        in: query
        name: from
        type: integer
      - description: 结束 unix 秒（默认当前）
        in: query
        name: to
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Overview'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 总览聚合
      tags:
      - 时序
  /plugins:
    get:
      description: 全部插件（含各版本元信息，不含二进制内容）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Plugin'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 插件仓库
      tags:
      - 插件
    post:
      consumes:
      - multipart/form-data
      description: 上传不可变插件版本（multipart：file + id/name/version）；服务端计算 SHA-256
      parameters:
      - description: 插件二进制
        in: formData
        name: file
        required: true
        type: file
      - description: 插件标识（稳定跨版本）
        in: formData
        name: id
        required: true
        type: string
      - description: 展示名
        in: formData
        name: name
        required: true
        type: string
      - description: 版本号（同 id 不可重复）
        in: formData
        name: version
        required: true
        type: string
      - description: 参数说明（前端表单用，JSON）
        in: formData
        name: args_schema
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Plugin'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 版本已存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 上传插件
      tags:
      - 插件
  /plugins/{id}:
    get:
      description: 某插件 id 的全部版本元信息
      parameters:
      - description: 插件标识
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Plugin'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 插件详情
      tags:
      - 插件
  /plugins/{id}/assign:
    post:
      consumes:
      - application/json
      description: 将某版本插件指派到指定 agent（覆盖旧指派）；agent 下次心跳拉取并执行
      parameters:
      - description: 插件标识
        in: path
        name: id
        required: true
        type: string
      - description: agent_id / version / args_json
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.assignPluginBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 插件版本不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 指派插件
      tags:
      - 插件
  /settings:
    get:
      description: 飞书 webhook 与 secret 是否存在（secret 不回显明文）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.settingsView'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 读取设置
      tags:
      - 设置
    put:
      consumes:
      - application/json
      description: 更新飞书 webhook；secret 为空表示保留原值
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.settingsView'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 保存设置
      tags:
      - 设置
  /settings/feishu-test:
    post:
      description: 按当前配置发送一条测试消息
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 未配置或发送失败
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 测试飞书机器人
      tags:
      - 设置
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.User'
            type: array
      security:
      - BearerAuth: []
      summary: 用户列表
      tags:
      - 用户
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名/初始密码/显示名
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.userBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: 用户名或密码不合规
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 用户名已存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 新建用户
      tags:
      - 用户
  /users/{id}:
    delete:
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 删自己 / 删最后一个用户
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: string
      - description: 显示名/角色
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.updateUserBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新用户
      tags:
      - 用户
...
securityDefinitions:
  BearerAuth:
    description: Web 面板 JWT（POST /api/auth/login 获取），格式：Bearer <jwt>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"